Move Core\Template Facade to Support\Facades namespace
